refactored listing.js and user.js routes
